Create an abstract ResourceProcessor class
The Functions ResourceProcessing & WasteProcessing are identical aparat from the selection of processes.
The logic can be pushed into a base class, ResourceProcessor, with the concreate classes just selecting the appropriate process list.

Changes:
- Create a ResourceProcessor class to unify the processing logic
- WasteProcessing & ResourceProcessing now use the common base class
- Console waste & process command use the common ResourceProcessor
- ConsoldeHelper has a new helper method to display the processes.
- Processes are displayed in a more compact format

Part of #739
